Wrap clay toy in iframe for production
