+++ toys/clay.html
@@ -52,7 +52,7 @@
       </div>
     </div>
     <script type="module">
-      import { bootstrapClayPage } from '../assets/js/toys/clay.ts';
+      import { bootstrapClayPage } from '../assets/js/toys/clay-toy.ts';
 
       bootstrapClayPage();
     </script>
